// src/generated/api/models/billSplitServiceApplicationFeaturesGroupsGetGroupGetGroupHandlerResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Bill Split Service API
 * API for bill splitting, Google login, payment tracking and administration.
 * OpenAPI spec version: v1
 */
import type { BillSplitServiceApplicationFeaturesGroupsGetGroupGetGroupHandlerBillItem } from './billSplitServiceApplicationFeaturesGroupsGetGroupGetGroupHandlerBillItem';
import type { BillSplitServiceApplicationFeaturesGroupsGetGroupGetGroupHandlerMemberItem } from './billSplitServiceApplicationFeaturesGroupsGetGroupGetGroupHandlerMemberItem';

export interface BillSplitServiceApplicationFeaturesGroupsGetGroupGetGroupHandlerResponse {
  groupId?: string;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  status?: string | null;
  isOwner?: boolean;
  /** @nullable */
  members?: BillSplitServiceApplicationFeaturesGroupsGetGroupGetGroupHandlerMemberItem[] | null;
  /** @nullable */
  bills?: BillSplitServiceApplicationFeaturesGroupsGetGroupGetGroupHandlerBillItem[] | null;
}