// src/generated/api/models/billSplitServiceApplicationFeaturesGroupsListGroupsListGroupsHandlerItem.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Bill Split Service API
 * API for bill splitting, Google login, payment tracking and administration.
 * OpenAPI spec version: v1
 */

export interface BillSplitServiceApplicationFeaturesGroupsListGroupsListGroupsHandlerItem {
  groupId?: string;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  role?: string | null;
  memberCount?: number;
  billCount?: number;
  /** @nullable */
  status?: string | null;
  updatedAtUtc?: string;
}